user profile pic update

+++ app/Http/Controllers/UserController.php
@@ -4,6 +4,10 @@ namespace App\Http\Controllers;
 
 use App\User;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\File;
+use Illuminate\Support\Facades\Hash;
+use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 
 class UserController extends Controller
@@ -20,29 +24,10 @@ class UserController extends Controller
      */
     public function index()
     {
-        $users = User::all();
-        return view('users.index', compact('users'));
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
-    public function create()
-    {
-        //
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
-    public function store(Request $request)
-    {
-        //
+        if(Auth::user()->role == 1 ) {
+            $users = User::all();
+            return view('users.index', compact('users'));
+        }
     }
 
     /**
@@ -65,8 +50,10 @@ class UserController extends Controller
      */
     public function edit($id)
     {
-        $user = User::findOrFail($id);
-        return view('users.edit', compact('user'));
+        if(Auth::user()->role == 1 ) {
+            $user = User::findOrFail($id);
+            return view('users.edit', compact('user'));
+        }
     }
 
     /**
@@ -79,22 +66,35 @@ class UserController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $this->validate($request, [
-            'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255',],
-            'role' => ['required'],
-            'password' => ['required', 'string', 'min:8', 'confirmed'],
-        ]);
+        if(Auth::user()->role == 1) {
+            $this->validate($request, [
+                'name' => ['required', 'string', 'max:255'],
+                'email' => ['required', 'string', 'email', 'max:255',],
+                'role' => ['required'],
+                'filename' => ['mimes:jpeg,bmp,png'],
+                'password' => ['required', 'string', 'min:8', 'confirmed'],
+            ]);
 
-        $user = User::findOrFail($id);
+            if ($request->hasfile('filename')) {
+                $cover = $request->file('filename');
+                $extension = $cover->getClientOriginalExtension();
 
-        $user->name = $request->input('name');
-        $user->email = $request->input('email');
-        $user->role = $request->input('role');
-        $user->password = $request->input('password');
-        $user->save();
+                Storage::disk('public')->put($cover->getFilename() . '.' . $extension, File::get($cover));
+            }
 
-        return redirect('users')->with('success','User updated successfully');
+            $user = User::findOrFail($id);
+
+            $user->name = $request->input('name');
+            $user->email = $request->input('email');
+            $user->role = $request->input('role');
+            $user->mime = isset($cover) ? $cover->getClientMimeType() : null;
+            $user->original_filename = isset($cover) ? $cover->getClientOriginalName() : null;
+            $user->filename = isset($cover) ? $cover->getFilename().'.'.$extension : null;
+            $user->password = Hash::make($request->input('password'));
+            $user->save();
+
+            return redirect('users')->with('success', 'User updated successfully');
+        }
     }
 
     /**
@@ -105,8 +105,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
-        $user->delete();
-        return redirect('users')->with('success','User deleted successfully');
+        if(Auth::user()->role == 1 ) {
+            $user = User::findOrFail($id);
+            $user->delete();
+            return redirect('users')->with('success', 'User deleted successfully');
+        }
     }
 }
